+++ test/parallel_tasks_test.dart
@@ -102,8 +102,10 @@ void main() {
           contains(RegExp(r'ERROR - Multiple stackTraces', multiLine: true)));
       expect(tasksOutput,
           contains(RegExp(r'\s*#[0|1]\s+_File.open.', multiLine: true)));
-      expect(tasksOutput,
-          contains(RegExp(r'\s*#[1|2]\s+_readNonExistingFile', multiLine: true)));
+      expect(
+          tasksOutput,
+          contains(
+              RegExp(r'\s*#[1|2]\s+_readNonExistingFile', multiLine: true)));
       expect(tasksOutput,
           contains(RegExp(r'\s*#[2|3]\s+throwError', multiLine: true)));
       expect(tasksOutput,