+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## done, waiting for next release
 
+# 0.20.0
+
 ### Dartle Core
 
 - fixed bug in `deleteAll` in which it deleted full directory, not respecting `fileExtensions` or `exclusions` filters.